Trace: pointer to char wasn't handled correctly as a trace param. Don't suspend the process on PANIC.

+++ src/plasma/trace/dbuffer.hpp
@@ -43,7 +43,7 @@ public:
     void flush();
 
 private:
-    Buffer *get_buffer(uint64_t generation);
+    Buffer *get_buffer(uint32_t generation);
 
     Buffer *_buffers;
     std::atomic<uint32_t> _generation;
